requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langchain==0.3.27
langchain-openai==0.3.7
langchain-ollama==0.2.0
langchain-community==0.3.27
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.4.0
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
ruff==0.3.4

# Search tools
tavily-python>=0.3.0
duckduckgo-search>=5.0
arxiv>=2.1.0
wikipedia>=1.4.0

# Web scraping
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
